Reorganize webpack configuration structure
