+++ packages/react-scripts/config/webpack.rules.js
@@ -1,0 +1,310 @@
+// @remove-on-eject-begin
+/**
+ * Copyright (c) 2015-present, Facebook, Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+// @remove-on-eject-end
+'use strict';
+
+const eslint = require('eslint');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const postcssNormalize = require('postcss-normalize');
+
+const paths = require(',/paths');
+const getCacheIdentifier = require('react-dev-utils/getCacheIdentifier');
+const getCSSModuleLocalIdent = require('react-dev-utils/getCSSModuleLocalIdent');
+
+// style files regexes
+const cssRegex = /\.css$/;
+const cssModuleRegex = /\.module\.css$/;
+const sassRegex = /\.(scss|sass)$/;
+const sassModuleRegex = /\.module\.(scss|sass)$/;
+
+module.exports = ({
+  isEnvDevelopment,
+  isEnvProduction,
+  shouldUseRelativeAssetPaths,
+  shouldUseSourceMap,
+  imageInlineSizeLimit,
+  mode,
+}) => {
+  // common function to get style loaders
+  const getStyleLoaders = (cssOptions, preProcessor) => {
+    const loaders = [
+      isEnvDevelopment && require.resolve('style-loader'),
+      isEnvProduction && {
+        loader: MiniCssExtractPlugin.loader,
+        options: shouldUseRelativeAssetPaths ? { publicPath: '../../' } : {},
+      },
+      {
+        loader: require.resolve('css-loader'),
+        options: cssOptions,
+      },
+      {
+        // Options for PostCSS as we reference these options twice
+        // Adds vendor prefixing based on your specified browser support in
+        // package.json
+        loader: require.resolve('postcss-loader'),
+        options: {
+          // Necessary for external CSS imports to work
+          // https://github.com/facebook/create-react-app/issues/2677
+          ident: 'postcss',
+          plugins: () => [
+            require('postcss-flexbugs-fixes'),
+            require('postcss-preset-env')({
+              autoprefixer: {
+                flexbox: 'no-2009',
+              },
+              stage: 3,
+            }),
+            // Adds PostCSS Normalize as the reset css with default options,
+            // so that it honors browserslist config in package.json
+            // which in turn let's users customize the target behavior as per their needs.
+            postcssNormalize(),
+          ],
+          sourceMap: isEnvProduction && shouldUseSourceMap,
+        },
+      },
+    ].filter(Boolean);
+    if (preProcessor) {
+      loaders.push({
+        loader: require.resolve(preProcessor),
+        options: {
+          sourceMap: isEnvProduction && shouldUseSourceMap,
+        },
+      });
+    }
+    return loaders;
+  };
+
+  return [
+    // Disable require.ensure as it's not a standard language feature.
+    { parser: { requireEnsure: false } },
+    // First, run the linter.
+    // It's important to do this before Babel processes the JS.
+    {
+      test: /\.(js|mjs|jsx|ts|tsx)$/,
+      enforce: 'pre',
+      use: [
+        {
+          options: {
+            formatter: require.resolve('react-dev-utils/eslintFormatter'),
+            eslintPath: require.resolve('eslint'),
+            // @remove-on-eject-begin
+            baseConfig: (() => {
+              const eslintCli = new eslint.CLIEngine();
+              let eslintConfig;
+              try {
+                eslintConfig = eslintCli.getConfigForFile(paths.appIndexJs);
+              } catch (e) {
+                // A config couldn't be found.
+              }
+
+              // We allow overriding the config, only if it extends our config
+              // (`extends` can be a string or array of strings).
+              if (
+                process.env.EXTEND_ESLINT &&
+                eslintConfig &&
+                eslintConfig.extends &&
+                eslintConfig.extends.includes('react-app')
+              ) {
+                return eslintConfig;
+              } else {
+                return {
+                  extends: [require.resolve('eslint-config-react-app')],
+                };
+              }
+            })(),
+            ignore: false,
+            useEslintrc: false,
+            // @remove-on-eject-end
+          },
+          loader: require.resolve('eslint-loader'),
+        },
+      ],
+      include: paths.appSrc,
+    },
+    {
+      // "oneOf" will traverse all following loaders until one will
+      // match the requirements. When no loader matches it will fall
+      // back to the "file" loader at the end of the loader list.
+      oneOf: [
+        // "url" loader works like "file" loader except that it embeds assets
+        // smaller than specified limit in bytes as data URLs to avoid requests.
+        // A missing `test` is equivalent to a match.
+        {
+          test: [/\.bmp$/, /\.gif$/, /\.jpe?g$/, /\.png$/],
+          loader: require.resolve('url-loader'),
+          options: {
+            limit: imageInlineSizeLimit,
+            name: 'static/media/[name].[hash:8].[ext]',
+          },
+        },
+        // Process application JS with Babel.
+        // The preset includes JSX, Flow, TypeScript, and some ESnext features.
+        {
+          test: /\.(js|mjs|jsx|ts|tsx)$/,
+          include: paths.appSrc,
+          loader: require.resolve('babel-loader'),
+          options: {
+            customize: require.resolve(
+              'babel-preset-react-app/webpack-overrides'
+            ),
+            // @remove-on-eject-begin
+            babelrc: false,
+            configFile: false,
+            presets: [require.resolve('babel-preset-react-app')],
+            // Make sure we have a unique cache identifier, erring on the
+            // side of caution.
+            // We remove this when the user ejects because the default
+            // is sane and uses Babel options. Instead of options, we use
+            // the react-scripts and babel-preset-react-app versions.
+            cacheIdentifier: getCacheIdentifier(mode, [
+              'babel-plugin-named-asset-import',
+              'babel-preset-react-app',
+              'react-dev-utils',
+              'react-scripts',
+            ]),
+            // @remove-on-eject-end
+            plugins: [
+              [
+                require.resolve('babel-plugin-named-asset-import'),
+                {
+                  loaderMap: {
+                    svg: {
+                      ReactComponent:
+                        '@svgr/webpack?-svgo,+titleProp,+ref![path]',
+                    },
+                  },
+                },
+              ],
+            ],
+            // This is a feature of `babel-loader` for webpack (not Babel itself).
+            // It enables caching results in ./node_modules/.cache/babel-loader/
+            // directory for faster rebuilds.
+            cacheDirectory: true,
+            cacheCompression: isEnvProduction,
+            compact: isEnvProduction,
+          },
+        },
+        // Process any JS outside of the app with Babel.
+        // Unlike the application JS, we only compile the standard ES features.
+        {
+          test: /\.(js|mjs)$/,
+          exclude: /@babel(?:\/|\\{1,2})runtime/,
+          loader: require.resolve('babel-loader'),
+          options: {
+            babelrc: false,
+            configFile: false,
+            compact: false,
+            presets: [
+              [
+                require.resolve('babel-preset-react-app/dependencies'),
+                { helpers: true },
+              ],
+            ],
+            cacheDirectory: true,
+            cacheCompression: isEnvProduction,
+            // @remove-on-eject-begin
+            cacheIdentifier: getCacheIdentifier(mode, [
+              'babel-plugin-named-asset-import',
+              'babel-preset-react-app',
+              'react-dev-utils',
+              'react-scripts',
+            ]),
+            // @remove-on-eject-end
+            // If an error happens in a package, it's possible to be
+            // because it was compiled. Thus, we don't want the browser
+            // debugger to show the original code. Instead, the code
+            // being evaluated would be much more helpful.
+            sourceMaps: false,
+          },
+        },
+        // "postcss" loader applies autoprefixer to our CSS.
+        // "css" loader resolves paths in CSS and adds assets as dependencies.
+        // "style" loader turns CSS into JS modules that inject <style> tags.
+        // In production, we use MiniCSSExtractPlugin to extract that CSS
+        // to a file, but in development "style" loader enables hot editing
+        // of CSS.
+        // By default we support CSS Modules with the extension .module.css
+        {
+          test: cssRegex,
+          exclude: cssModuleRegex,
+          use: getStyleLoaders({
+            importLoaders: 1,
+            sourceMap: isEnvProduction && shouldUseSourceMap,
+          }),
+          // Don't consider CSS imports dead code even if the
+          // containing package claims to have no side effects.
+          // Remove this when webpack adds a warning or an error for this.
+          // See https://github.com/webpack/webpack/issues/6571
+          sideEffects: true,
+        },
+        // Adds support for CSS Modules (https://github.com/css-modules/css-modules)
+        // using the extension .module.css
+        {
+          test: cssModuleRegex,
+          use: getStyleLoaders({
+            importLoaders: 1,
+            sourceMap: isEnvProduction && shouldUseSourceMap,
+            modules: true,
+            getLocalIdent: getCSSModuleLocalIdent,
+          }),
+        },
+        // Opt-in support for SASS (using .scss or .sass extensions).
+        // By default we support SASS Modules with the
+        // extensions .module.scss or .module.sass
+        {
+          test: sassRegex,
+          exclude: sassModuleRegex,
+          use: getStyleLoaders(
+            {
+              importLoaders: 2,
+              sourceMap: isEnvProduction && shouldUseSourceMap,
+            },
+            'sass-loader'
+          ),
+          // Don't consider CSS imports dead code even if the
+          // containing package claims to have no side effects.
+          // Remove this when webpack adds a warning or an error for this.
+          // See https://github.com/webpack/webpack/issues/6571
+          sideEffects: true,
+        },
+        // Adds support for CSS Modules, but using SASS
+        // using the extension .module.scss or .module.sass
+        {
+          test: sassModuleRegex,
+          use: getStyleLoaders(
+            {
+              importLoaders: 2,
+              sourceMap: isEnvProduction && shouldUseSourceMap,
+              modules: true,
+              getLocalIdent: getCSSModuleLocalIdent,
+            },
+            'sass-loader'
+          ),
+        },
+        // "file" loader makes sure those assets get served by WebpackDevServer.
+        // When you `import` an asset, you get its (virtual) filename.
+        // In production, they would get copied to the `build` folder.
+        // This loader doesn't use a "test" so it will catch all modules
+        // that fall through the other loaders.
+        {
+          loader: require.resolve('file-loader'),
+          // Exclude `js` files to keep "css" loader working as it injects
+          // its runtime that would otherwise be processed through "file" loader.
+          // Also exclude `html` and `json` extensions so they get processed
+          // by webpacks internal loaders.
+          exclude: [/\.(js|mjs|jsx|ts|tsx)$/, /\.html$/, /\.json$/],
+          options: {
+            name: 'static/media/[name].[hash:8].[ext]',
+          },
+        },
+        // ** STOP ** Are you adding a new loader?
+        // Make sure to add the new loader(s) before the "file" loader.
+      ],
+    },
+  ];
+};
